pyproject.toml
[tool.poetry]
name = "coursework7"
version = "0.1.0"
description = ""
authors = ["ika <mamina747@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
pytz = "^2024.2"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
